refactor: the generator module
