+++ hybridse/src/vm/catalog_wrapper.h
@@ -16,21 +16,17 @@
 
 #ifndef HYBRIDSE_SRC_VM_CATALOG_WRAPPER_H_
 #define HYBRIDSE_SRC_VM_CATALOG_WRAPPER_H_
+
 #include <memory>
 #include <string>
 #include <utility>
+
 #include "vm/catalog.h"
+#include "vm/generator.h"
+
 namespace hybridse {
 namespace vm {
 
-class ProjectFun {
- public:
-    virtual Row operator()(const Row& row, const Row& parameter) const = 0;
-};
-class PredicateFun {
- public:
-    virtual bool operator()(const Row& row, const Row& parameter) const = 0;
-};
 class IteratorProjectWrapper : public RowIterator {
  public:
     IteratorProjectWrapper(std::unique_ptr<RowIterator>&& iter, const Row& parameter, const ProjectFun* fun)
@@ -564,6 +560,151 @@ class RowCombineWrapper : public RowHandler {
     size_t right_slices_;
     Row value_;
     const ProjectFun* fun_;
+};
+
+class LazyLastJoinIterator : public RowIterator {
+ public:
+    LazyLastJoinIterator(std::unique_ptr<RowIterator>&& left, std::shared_ptr<PartitionHandler> right, const Row& param,
+                         std::shared_ptr<JoinGenerator> join)
+        : left_it_(std::move(left)), right_(right), parameter_(param), join_(join) {}
+
+    ~LazyLastJoinIterator() override {}
+
+    bool Valid() const override {
+        return left_it_ && left_it_->Valid();
+    }
+    void Next() override {
+        left_it_->Next();
+    }
+    const uint64_t& GetKey() const override {
+        return left_it_->GetKey();
+    }
+    const Row& GetValue() override {
+        value_ = join_->RowLastJoin(left_it_->GetValue(), right_, parameter_);
+        return value_;
+    }
+    bool IsSeekable() const override {
+        return true;
+    };
+
+    void Seek(const uint64_t& key) override {
+        left_it_->Seek(key);
+    }
+
+    void SeekToFirst() override {
+        left_it_->SeekToFirst();
+    };
+
+ private:
+    std::unique_ptr<RowIterator> left_it_;
+    std::shared_ptr<PartitionHandler> right_;
+    const Row& parameter_;
+    std::shared_ptr<JoinGenerator> join_;
+
+    Row value_;
+};
+
+class LazyLastJoinPartitionHandler final : public PartitionHandler {
+ public:
+    LazyLastJoinPartitionHandler(std::shared_ptr<PartitionHandler> left, std::shared_ptr<PartitionHandler> right,
+                                const Row& param, std::shared_ptr<JoinGenerator> join);
+    ~LazyLastJoinPartitionHandler() override {}
+
+    // NOTE: only support get segement by key from left source
+    std::shared_ptr<TableHandler> GetSegment(const std::string& key) override;
+
+    std::unique_ptr<WindowIterator> GetWindowIterator() override {
+        return left_->GetWindowIterator();
+    }
+
+    const std::string GetHandlerTypeName() override {
+        return "LazyLastJoinPartitionHandler";
+    }
+    std::unique_ptr<RowIterator> GetIterator() override {
+        auto iter = left_->GetIterator();
+        if (!iter) {
+            return std::unique_ptr<RowIterator>();
+        }
+        return std::unique_ptr<RowIterator>(new LazyLastJoinIterator(std::move(iter), right_, parameter_, join_));
+    }
+    const Types& GetTypes() override { return left_->GetTypes(); }
+    const IndexHint& GetIndex() override { return left_->GetIndex(); }
+
+    const Schema* GetSchema() override { return nullptr; }
+    const std::string& GetName() override { return name_; }
+    const std::string& GetDatabase() override { return db_; }
+
+    base::ConstIterator<uint64_t, Row>* GetRawIterator() override {
+        return nullptr;
+    }
+
+ private:
+    std::shared_ptr<PartitionHandler> left_;
+    std::shared_ptr<PartitionHandler> right_;
+    const Row& parameter_;
+    std::shared_ptr<JoinGenerator> join_;
+
+    std::string name_ = "";
+    std::string db_ = "";
+};
+
+class LazyLastJoinTableHandler final : public TableHandler {
+ public:
+    LazyLastJoinTableHandler(std::shared_ptr<TableHandler> left, std::shared_ptr<PartitionHandler> right,
+                                const Row& param, std::shared_ptr<JoinGenerator> join);
+    ~LazyLastJoinTableHandler() override {}
+
+    std::unique_ptr<RowIterator> GetIterator() override {
+        auto iter = left_->GetIterator();
+        if (!iter) {
+            return std::unique_ptr<RowIterator>();
+        }
+
+        return std::unique_ptr<RowIterator>(new LazyLastJoinIterator(std::move(iter), right_, parameter_, join_));
+    }
+
+    const Types& GetTypes() override { return left_->GetTypes(); }
+    const IndexHint& GetIndex() override { return left_->GetIndex(); }
+
+    std::unique_ptr<WindowIterator> GetWindowIterator(
+        const std::string& idx_name) override {
+        return nullptr;
+    }
+
+    const Schema* GetSchema() override { return nullptr; }
+    const std::string& GetName() override { return name_; }
+    const std::string& GetDatabase() override { return db_; }
+
+    base::ConstIterator<uint64_t, Row>* GetRawIterator() override {
+        return nullptr;
+    }
+
+    Row At(uint64_t pos) override {
+        return value_;
+    }
+
+    const uint64_t GetCount() override { return left_->GetCount(); }
+
+    std::shared_ptr<PartitionHandler> GetPartition(const std::string& index_name) override {
+        return std::shared_ptr<PartitionHandler>(
+            new LazyLastJoinPartitionHandler(left_->GetPartition(index_name), right_, parameter_, join_));
+    }
+
+    const OrderType GetOrderType() const override { return left_->GetOrderType(); }
+
+    const std::string GetHandlerTypeName() override {
+        return "LazyLastJoinTableHandler";
+    }
+
+ private:
+    std::shared_ptr<TableHandler> left_;
+    std::shared_ptr<PartitionHandler> right_;
+    const Row& parameter_;
+    std::shared_ptr<JoinGenerator> join_;
+
+    Row value_;
+    std::string name_ = "";
+    std::string db_ = "";
 };
 
 }  // namespace vm
